MAS/app/agents/requirements.txt
pandas
yfinance
numpy
matplotlib
torch
torchvision
scikit-learn
scikit-fuzzy
tweepy
configparser
google-cloud-language
